backend/requirements.txt
fastapi
uvicorn
pymupdf
scikit-learn
requests
tiktoken
python-dotenv
pydantic
pydantic-settings
python-multipart
